+++ src/main/java/com/project/cheerha/common/repository/KeyValueCommandRepository.java
@@ -1,4 +1,4 @@
-package com.project.cheerha.common.repository.keyvalue;
+package com.project.cheerha.common.repository;
 
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
