+++ src/main/java/com/project/cheerha/domain/auth/service/BlackListService.java
@@ -1,8 +1,8 @@
 package com.project.cheerha.domain.auth.service;
 
 import com.project.cheerha.common.properties.JwtSecurityProperties;
-import com.project.cheerha.common.repository.keyvalue.KeyValueCommandRepository;
-import com.project.cheerha.common.repository.keyvalue.KeyValueQueryRepository;
+import com.project.cheerha.common.repository.KeyValueCommandRepository;
+import com.project.cheerha.common.repository.KeyValueQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
