+++ src/test/java/com/project/cheerha/redis/emailtoken/SaveTokenTest.java
@@ -2,7 +2,7 @@ package com.project.cheerha.redis.emailtoken;
 
 import com.project.cheerha.common.exception.client.BadRequestException;
 import com.project.cheerha.common.exception.client.ClientErrorCode;
-import com.project.cheerha.common.repository.keyvalue.KeyValueCommandRepository;
+import com.project.cheerha.common.repository.KeyValueCommandRepository;
 import com.project.cheerha.domain.user.service.CheckDailyEmailCount;
 import com.project.cheerha.domain.user.service.EmailTokenService;
 import org.junit.jupiter.api.Test;
