+++ src/test/java/com/project/cheerha/redis/emailtoken/VerifyEmailTokenTest.java
@@ -1,8 +1,8 @@
 package com.project.cheerha.redis.emailtoken;
 
 import com.project.cheerha.common.exception.client.BadRequestException;
-import com.project.cheerha.common.repository.keyvalue.KeyValueCommandRepository;
-import com.project.cheerha.common.repository.keyvalue.KeyValueQueryRepository;
+import com.project.cheerha.common.repository.KeyValueCommandRepository;
+import com.project.cheerha.common.repository.KeyValueQueryRepository;
 import com.project.cheerha.domain.user.service.EmailTokenService;
 import com.project.cheerha.domain.user.service.VerificationFailCount;
 import org.junit.jupiter.api.Test;
